Refactoring opening of DB to config/helpers.js
